refactor: Extract page components into separate files
- Extract Header component from page.tsx
- Extract HeroSection component from page.tsx
- Extract FeaturesSection component from page.tsx
- Extract HowItWorksSection component from page.tsx
- Extract RoadmapSection component from page.tsx
- Extract CTASection component from page.tsx
- Extract Footer component from page.tsx
- Update page.tsx to import and use new components

This improves code organization and maintainability by following
the component-based architecture pattern already established with
the FAQ section.
